+++ src/util/spanparsing.h
@@ -6,10 +6,9 @@
 #define BITCOIN_UTIL_SPANPARSING_H
 
 #include <span.h>
+#include <util/string.h>
 
 #include <string>
-#include <string_view>
-#include <vector>
 
 namespace spanparsing {
 
@@ -37,41 +36,11 @@ bool Func(const std::string& str, Span<const char>& sp);
  */
 Span<const char> Expr(Span<const char>& sp);
 
-/** Split a string on any char found in separators, returning a vector.
- *
- * If sep does not occur in sp, a singleton with the entirety of sp is returned.
- *
- * Note that this function does not care about braces, so splitting
- * "foo(bar(1),2),3) on ',' will return {"foo(bar(1)", "2)", "3)"}.
- */
-template <typename T = Span<const char>>
-std::vector<T> Split(const Span<const char>& sp, std::string_view separators)
+/** Split alias for backwards compatibility */
+template <typename... Args>
+auto Split(Args&&... args)
 {
-    std::vector<T> ret;
-    auto it = sp.begin();
-    auto start = it;
-    while (it != sp.end()) {
-        if (separators.find(*it) != std::string::npos) {
-            ret.emplace_back(start, it);
-            start = it + 1;
-        }
-        ++it;
-    }
-    ret.emplace_back(start, it);
-    return ret;
-}
-
-/** Split a string on every instance of sep, returning a vector.
- *
- * If sep does not occur in sp, a singleton with the entirety of sp is returned.
- *
- * Note that this function does not care about braces, so splitting
- * "foo(bar(1),2),3) on ',' will return {"foo(bar(1)", "2)", "3)"}.
- */
-template <typename T = Span<const char>>
-std::vector<T> Split(const Span<const char>& sp, char sep)
-{
-    return Split<T>(sp, std::string_view{&sep, 1});
+    return ::Split(std::forward<Args>(args)...);
 }
 
 } // namespace spanparsing
